chore(back): remove abstract file store service
I don't think we'll use anything but s3 compatible storage
