+++ apps/backend/src/app/modules/filestore/file-store.service.spec.ts
@@ -2,18 +2,18 @@ import { createHash } from 'node:crypto';
 import { Test, TestingModule } from '@nestjs/testing';
 import { ConfigService } from '@nestjs/config';
 import { mockDeep } from 'jest-mock-extended';
-import { FileStoreS3Service } from './file-store-s3.service';
+import { FileStoreService } from './file-store.service';
 import { LargeFileStoreService } from './large-file-store.service';
 import { arrayFrom } from '@momentum/util-fn';
 
 describe('FileStoreS3Service', () => {
-  let service: FileStoreS3Service;
+  let service: FileStoreService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         {
-          provide: FileStoreS3Service,
+          provide: FileStoreService,
           useClass: LargeFileStoreService
         }
       ]
@@ -26,7 +26,7 @@ describe('FileStoreS3Service', () => {
       })
       .compile();
 
-    service = module.get<FileStoreS3Service>(FileStoreS3Service);
+    service = module.get<FileStoreService>(FileStoreService);
   });
 
   it('should be defined', () => {
